! Add url checking before fetching title
